Move VoteButtons to its own file
