+++ keg-client/src/Taps/BeerBlock.tsx
@@ -1,10 +1,9 @@
 import React from 'react';
-import clsx from 'clsx';
 import { makeStyles } from '@material-ui/core/styles';
 import { Beer } from '../ServerModels';
+import { VoteButtons } from './VoteButtons';
 import Box from '@material-ui/core/Box';
 import Button from '@material-ui/core/Button';
-import ButtonGroup from '@material-ui/core/ButtonGroup';
 import Card from '@material-ui/core/Card';
 import CardContent from '@material-ui/core/CardContent';
 import Chip from '@material-ui/core/Chip';
@@ -20,21 +19,6 @@ import './BeerBlock.css';
 
 const useStyles = makeStyles(theme => ({
   card: {},
-  button: {
-    margin: theme.spacing(0.25),
-  },
-  upVoteButton: {
-    backgroundColor: '#40BF40',
-    '&:hover': {
-      background: '#2E892E',
-    },
-  },
-  downVoteButton: {
-    backgroundColor: '#BF4040',
-    '&:hover': {
-      background: '#913131',
-    },
-  },
   leftIcon: {
     marginRight: theme.spacing(1),
   },
@@ -156,50 +140,5 @@ function DetailDialog(props: DetailDialogProps) {
         </Button>
       </DialogActions>
     </Dialog>
-  );
-}
-
-export interface VoteButtonsProps {
-  onClick: (isUpVote: boolean) => void;
-  upVotes: number;
-  downVotes: number;
-}
-
-function VoteButtons(props: VoteButtonsProps) {
-  const classes = useStyles();
-
-  const handleUpVote = () => {
-    props.onClick(true);
-  };
-
-  const handleDownVote = () => {
-    props.onClick(false);
-  };
-
-  return (
-    <React.Fragment>
-      <ButtonGroup>
-        <Button
-          size="small"
-          color="primary"
-          variant="contained"
-          className={clsx(classes.button, classes.upVoteButton)}
-          onClick={handleUpVote}
-        >
-          <Icon className={clsx(classes.leftIcon, classes.iconSmall)}>thumb_up</Icon>
-          {props.upVotes}
-        </Button>
-        <Button
-          size="small"
-          color="secondary"
-          variant="contained"
-          className={clsx(classes.button, classes.downVoteButton)}
-          onClick={handleDownVote}
-        >
-          <Icon className={clsx(classes.leftIcon, classes.iconSmall)}>thumb_down</Icon>
-          {props.downVotes}
-        </Button>
-      </ButtonGroup>
-    </React.Fragment>
   );
 }
